// src/enum/return-items-status.enum.ts
export enum ReturnItemStatus {
  RETURN_REQUESTED = 'RETURN_REQUESTED',
  RETURN_ACCEPTED = 'RETURN_ACCEPTED',
  RETURN_IN_TRANSIT='RETURN_IN_TRANSIT',
  RETURN_RECEIVED = 'RETURN_RECEIVED',
  WAITING_APPROVAL = 'WAITING_APPROVAL',
  SPLIT = 'SPLIT',
  REFUNDED = 'REFUNDED',
  REPLACED = 'REPLACED',
  COMPLETED = 'COMPLETED',
}